Add FloatMath wrapper class
The version of PhotoView that we are using references the
android.util.FloatMath class, which was removed in API 23. We can either
update PhotoView, or copy in the FloatMath class from AOSP, which was
simply a compatibility wrapper for some methods in the java.lang.Math
class. I opted to do the latter, since I don't have the resources to
properly test if there are any regressions in PhotoView that could
affect us, like this one:
https://github.com/chrisbanes/PhotoView/issues/333